Fixes: Issue #1
Problem: cluster.py contains logic bug when checking for local device
name or address

Analysis: Found that an if statement was checking whether a string was
true or false by simply doing an if string_var. This only returns false
if the string is empty. Not when the string contains the word "False".
Fixed it to check for truth by doing if string_var == "true".

Unit/Robot tests: Tested deployment on an onboard server and ran the
cluster_generic.py command.