Create a base class for interfaces and refactor nat.py to use it
Issues:
Fixes #16
Fixes #17

Problem:
There is a lot of duplicate code in the various interface modules
that can be refactored into an inheritable base class for
an interface.

Analysis:
Create an Interfaces() base class that has some primitive private
functions as well as some basic common functions.  Refactor the
nat.py interface to use the new base class as an example
and create unit tests for the interfaces module.

Tests:
Created and ran unit tests for the Interfaces() class.