Fixes: #20 Add make targets to build rpms/debs using Docker
Problem:
In order to build packages for a specific OS the developer needs
to be on that OS.  This commit adds new make targets that rely on
docker containers to build the packages.

Analysis:
Add Dockerfiles and modify Makefile to introduce two need targets
docker_debs and docker_rpms.  These will create the docker images
that can build a package for RedHat/Centos 7 or Ubuntu Trusty

Unit/Robot tests:
Build targets and install.