Fix the f5-common-python packaging to correspond to the new repo structure
Issues:
Fixes #70

Problem:
The setup.py will not work with the new structure of the code.

Analysis:
Modify setup.py to use find_packages so that any extentions will be
automatically included.  Cleanup the versioning and release varibles
make sure that they are packaged as part of the python distributable.

Update the setup.py to include a install_requires and added the f5-icontrol-rest
package.

Tests:
Build and install.