Handle Exclusive Object Attributes
Issues:
Fixes #144

Problem:
There are object attributes that cannot be set at the same time.
For example in the f5.bigip.net.vlan.interfaces object you cannot
have both the tagged and untagged attribute set because when you
do an update to the BIGIP it throws an error. We need to handle
this for users by removing any of the attributes that are
exclusive in their relationship.

Analysis:
* Added a mixin that can be used by objects that require this
  functionality.
* The mixin overrides the `__setattr__` for the class and
  if the attribute being set is in one of the sets in
  `self._meta_data['exclusive_attributes']` it removes the other
  members of that set prior to assigning the value to the attr.

**NOTE** You must call `super(CLASSNAME, self).__setattr__ at the
end of the mixin to be sure that you actually set the attribute
like you wanted to.

Tests: