Remove the BigIQ module from repo #152
Issues:
Fixes #152

Problem:
We do not support BigIQ and the code in the repo is stale. I also checked the previous version of f5-openstack-lbaasv1 and it doesn't seem to use this code. If we want to have BigIQ support we will need to do this as an additional project.

Analysis:
* Remove the `f5/bigiq/ folder and any files in it
* Remove the references to BigIQ in the README
* Remove the references to BigIQ in the Makefile

Tests:
* Flake8
* Unit