Remove the SOAP api code from the repo (pycontrol) #154
Issues:
Fixes #154

Problem:
The current library has a pycontrol module in it that manages a connection to the SOAP interface of the BigIP. After talking with @jgruber he sees no reason for us to keep it given the new object/manager direction for the library. We should be able to accomplish everything we need to with the REST API.

We should remove this code and if we need a SOAP session manager we should create an f5-icontrol-soap library like the f5-icontrol-rest library.

I did some looking around at the library and it doesn't appear that anything really uses this except for SSL certificates, but after discussion we can do this with the REST API and a SSLManager to be written in the future.

Analysis:
* Remove the module files
* Removed the code from the __init__.py of bigip
* The SSL code is the only code that uses this that I could find.

Tests:
* flake8
* Unit tests