Refactor iapp module into application
Issues: Issue #76

Problem: The iapp module needs to be refactored

Analysis: Implemented the application module which uses the new
OrganizingCollection, Collection, and Resource base classes to structure
BigIP resources according to the uri structure. The application module
contains a Template and Service resource as well as their respective
collections. It also contains an overarching Application
OrganizingCollection.

Tests: Some sys unit tests are failing. Will rebase new changes on top
of this to see if that fixes it.