Further fixups from code review comments
Issues: PR #175

Problem: The build_uri function is unnecessary and the uri needs to be
populated from the selfLink

Analysis: The build_uri function in the application module was leftover
from old code. It has been removed, as the uri needs to be built once,
just before the object is loaded. After that, the selfLink is used to
populate the meta_data['uri'] attribute properly.

Tests: All tests pass