Make load and create in concrete classes less susceptible to code drift.
Issues: PR #175

Problem: The _create and _load methods in the concrete Service class in
Application should use as much of the parent's code as possible,
minimizing duplicate code and possibility for drift.

Analysis: _create calls to its parent _create initially, and if an
expected exception is thrown in the parent, the child's _create calls
the overriden _load. This is coupled with some changes to resource.py to
ensure maximal resuse of code related to these two operations.

Tests: All unit and functional tests pass