Fixups for pull request #183, based on comments by @zancas
Issues: PR #183

Problem: Test interactions are occurring in this suite and an additional
test is needed to show it.

Analysis: The module level dictionaries were being manipulated in a
fixture and this is a side-effect behavior of the fixture, as the
dictionaries remain modified after the fixture has been torn down. Using
a copy() call when the dictionaries are used in the fixture. Also
created an additional set of two tests to show that this is happening.
Removed one of the tg tests in TestServiceUpdate because it was patching
both the create and update response from the BigIP and was testing very
little of application code.

Tests: All tests pass, 100% coverage