Implement OC, Coll, Resource from temp generation.
Issues:
Fixes: https://github.com/F5Networks/f5-common-python/issues/218

Problem: There's a lot of boilerplate involved in creating new concrete
subclasses.  This engine provides a tool to automatically generate that
boilerplate.

Analysis: The engine requires templates and JSON-config data as inputs
it produces Python boilerplate as a string as output.

Tests: These will be added in other stories.