Create an exists() method for resource objects. #226
Issues:
Fixes #226

Problem:
A very common thing for the projects using this SDK to do is to check
for the existence of an object. While this can be done by doing
something like using try/except to catch HTTPErrors on create and
then looking and the status code for 404 we can hide a lot of that
detail from the users with an exists() method.

Analysis:
* Added `exists()` method to the `Resource` base class.
* Added unit tests for 100% coverage of the new code
* Added a functional test to the nodes test to test on a live system.
* Added intersphinx config to link to the requests library.

Tests:
* Ran unit tests
* Ran functional tests
* Ran flake8 tests