Iapp parser must ignore curly braces inside double quotes
Issues: Issue #235

Problem: Curly braces are used as section boundaries in TCL and in iapp
templates. Yet, there is a case where these must be ignored, and that is
if they are inside a double-quoted string. We must ignore these curly
braces only for now.

Analysis: When searching for the end of an iapp template section, we are
looking for a closing curly brace '}', so to ignore the braces in
double-quoted strings, we must note when we are in a double-quoted
string and when we are not, that way we can then choose whether to
ingore the brace or not. This is how this code is implemented.

Tests: Added new unit test. All tests pass with 100% coverage