If an object already has a _meta_data['uri']
create and load will no longer talk to the device
Fixes #237

Problem: A loaded/created object shouldn't talk to the device.

Analysis: Telling the device to create a resource that you, the
client object, are already aware of, doesn't make sense and wastes
resources.

Tests: Updated Unittests, see the diff.