Application Services Service must override exists() to build proper uri
Issues: Issue #242

Problem: Override resource's exists() in application services service to
build the uri for exists

Analysis: Building the uri was made a method and the new exists method
uses this.

Tests: New unit tests were added to achieve 100% coverage and an
additional check was added in functional tests to determine existence of
resources.