Install via PIP fails compiling SSL #265
Issues:
Fixes #265

Problem:
We need to remove the dependency on the pyopenssl package because
this requires users to have a lot of outside dependency that are
not always available when installing the package. This was used
in the f5/bigip/ltm/ssl.py file that has not been refactored. We
need to remove this old code and the dependancy on the pyopenssl
package and figure out a cleaner way to do this if and when we
refactor that ssl file.

Analysis:
* Removed the ssl.py file, it will be refactored via issue #49
* Removed the pyopenssl from the requirements.*.txt files
* Removed the pyopenssl from the setup.py

Tests:
* Unit tests
* Functional tests
* Flake8
* Built the pypi package