Make requirements have SingleSourceofTruth
Issues:
Fixes #290

Problem: We had old requirements listed, and more than 1 source of truth.

Analysis: Using the pattern recommended here:

https://caremad.io/2013/07/setup-vs-requirement/

we've enforced a single source of truth, and removed obsolete requirements.

For general use of the f5-sdk library, add strings to the "install_requires"
parameter to the setup method in "setup.py".   For hacking on the codebase,
testing, etc. add packages and versions to the end of "requirements.txt".

NOTE:   The "-e ." line in the requirements.txt ensures that "pip install -r"
-ing that file will pick up the "install_requires" packages.

Tests: By hand verification of correct installation.