Need support for POSTing command to bigip.cm #272
Issues:
Fixes #272

Problem:
Need to be able to sync the configuration of device-groups via
the SDK.  This requires a special POST command to run a
command via the REST API with some options.

Analysis:
* Added a sync method to the `Cm` class that takes a device-group
name and exucutes the run command.
* Added a sync method to the `Device_Group` class as well that
calls the `Cm` container's sync method with it's name.
* Added tests for the cm.sync() and cm.device_groups.sync()

Tests:
* Built and reviewed docs
* Unit tests
* Functional tests