Adds support for connecting to BIG-IPs on off-numbered ports
Issues:
Fixes #347

Problem:
The f5-sdk makes the assumption that all BIG-IPs are accessed on
port 443. While this may be true in most cases, there are cases
where the BIG-IP's web ui may be accessed through a port-forward
and therefore the f5-sdk would need to specify the forwarded port
instead of assuming the default.

Analysis:
I added a port argument to the BigIP constructor. By default this
value is the standard port that the web ui is accessed on; 443.

Whether the port is specified or not, the string formatting performed
for the self._meta_data['uri'] value will include the port value.
The new format of the 'uri' value will resemble the following

    https://FakeHostName:443/mgmt/tm

If a port is specified, for instance 10443, it would then look like

    https://FakeHostName:10443/mgmt/tm

This change allows BIG-IPs to be accessed on off-numbered ports

Tests:

  * FakeBigIPWithPort
  * test_meta_data_uri