Add the HTTPD API to the SDK
Issues:
Fixes #372

Problem:
The issue is that the httpd API was not available in the SDK

Analysis:
This change adds the httpd api to the sdk. This API is scattered
throughout the Web UI, but maps neatly to the tmsh api. With this
API you can change various behaviors associated with the Web UI.

Tests:
  * f5/bigip/sys/test_httpd.py
  * test/functional/sys/test_httpd.py