Sync to group in cm failing to sync
Issues:
Fixes #396

Problem:
This is failing silently to sync the BigIP to group.

Analysis:
I have implemented a working example of syncing to and from group, yet
the caveat is that there is no good way to determine whether a group (or
device) has been synced. This is as far as I know. I perused each uri
path for the device_group and for the device itself, and came up with
nothing. Nevertheless, the tests pass.

Tests: All tests pass