Refactor clustering as part of review with @zancas
Issues:
Fixes #400

Problem:
Clustering is needed for creating a device service group cluster of
multiple bigip devices

Analysis:
This is part of a refactor from a code review that happened offline. It
was suggested to refactor the external interface that previously
existed with the Cluster class into ClusterManager, TrustDomain, and
DeviceGroup classes, which maintain state for their respective
operations and data. The interfaces presented by each of the classes
above are as uniform as possible, and allow for the ClusterManager to
manage the DeviceGroup and TrustDomain with ease.

Tests:
All previous unit tests and functional tests have been refactored or
rewritten in some way. All tests are currently passing.