Fixing opt-in for cluster tests and trust domain sync
Issues:
Fixes #456 and #460

Problem:
Developers must be able to opt-in to the cluster tests, since they
require a set of at least four bigip devices.

Also, a trust domain is implemented as a device group on the bigip
devices, and we must be able to sync that group when we add members to
the trust domain.

Analysis:
Used pytest.mark.skip to skip cluster tests unless a user has set a key
of 'run_cluster_tests' and a value of JSON/YAML truthiness.

The trust domain now instantiates a device group manager to sync to that
group when adding devices to the trust domain. This is a part of the
validation process in the device group and the trust domain.

Tests:
All unit and functional tests pass