Support two member cluster only in first clustering release
Issues:
Fixes #471

Problem:
In order to minimize feature creep and the ability for users to shoot
themselves in the foot, I would like to limit clustering support to
handle only two devices. This will be reflected in the code with tests
and documentation.

Analysis:
Modified the code to only allow up to four devices to be clustered, yet the
functional tests are there for up to two devices. Four devices is not yet
supported. This will be made clear in the customer-facing documentation.
It is already clear in the code documentation.

Tests:
All clustering tests pass against 11.6 and 12.0 VE images