Fixes the skipping of ltm pool tests
Issues:
Fixes #508

Problem:
The pytest.skip decorator was being used in a method of one of the test classes
and this was causing all of the tests in the file to be skipped.

Analysis:
The correct syntax for this is to use "pytest.mark.skip", but I took a look
and it turns out that these tests work just fine. So I removed the skip
entirely

Tests:
  * test/functional/tm/ltm/test_pool.py