Adds coveralls badge to repo
Issues:
Fixes #547

Problem:
The f5-sdk codebase can run coverage tests, but there is no easy reporting
of these results.

Analysis:
The Coveralls.io service has been setup for the f5-common-python test. This
change to the README is to add the badge for coveralls so that developers
can easily see what the coverage tests currently result in

Tests:
None