During functional testing in 11.5.4, requests_params with filters fail
Issues:
Fixes #559

Problem:
We've concluded that this occurs due to the way the 11.5.4 device fails
to interpret the encoded dollar sign (%24). We can work around this by
preventing requests from encoding our requests parameters.

Analysis:
This is a test fixup, but we'd like to encourage users of the sdk to
make these types of OData queries in a way that is compatible with >
11.5.4 versions. So there's a new document that details how to perform
these types of queries that is comptabile with 11.5.4 and greater.

Tests:
The requests params functional tests now pass and I've added one test to
check a select call.