During functional testing against 11.5.4, test_create_min_args fails in test_user.py suite
Issues:
Fixes #562

Problem:
The test in test/functional/auth/test_user fails due to the fact that in
11.5.4, the field partitionAccess is returned as a unicode, yet in 11.6
and greater, it is a dict. This just requires a test fixup for now.

Analysis:
Created a separate test for 11.5.4 and lower versions. The original
failing test is now skipped if the release version given to pytest is
lower than 11.6.0.

Tests:
Functional tests for test_user.py now pass