During functional testing against 11.5.4, test/functional/tm/ltm/test_policy.py tests produce errors
Issues:
Fixes #567

Problem:
During test teardown, an attempt is made to delete the policies that are
not a part of the built-ins. The built-ins have changed between versions
of TMOS. We should snapshot the get_collection call and use that
pre-test snapshot to delete all those policies created by the tests.

Analysis:
Implemented a module-level pytest fixture which snapshots the device
before and after the test. The diff between those snapshots is used to
remove objects created by the tests.

Tests:
test_policy tests now pass against 11.5.4 and 11.6.0