During functional testing in 11.5.4 device goes to forced offline and does not come back online
Issues:
Fixes #553

Problem:
If the tests failed in the middle of bringing the device to offline, the
device remains in that state. There are also sleeps in these tests,
which should be removed.

Analysis:
Implemented a pollster to check what state the device is in and created
a teardown that assures the device is active before leaving the test.

Tests:
All these tests pass against 11.5.4 and 11.6.0