During functional testing in 11.5.4 in ltm/profile some tests fail at setup
Issues:
Fixes #555

Problem:
We have 8 test failures against 11.5.4 and all but one have the same
signature: Found unexpected json pair at module

Analysis:
After further inspection, it's simply that these collections do not
exist on the 11.5.4 device. Added minimum tmos version to these
collections, so they will not be instantiated by a user. Added some
tests to ensure that the correct exception is raised if a user does try
to instantiate one of these collections.

Tests:
All profile tests are passing on 11.5.4 and 11.6.0