During functional testing against 12.1.0, sys/application test fails and errors out
Issues:
Fixes #580

Problem:
This is a bug in application.py where we expected the call to create an
application service to raise an HTTPError, which was happening in
versions prior to 12.1.0. This is now fixed on the BIG-IP and when
testing against 12.1.0, it revealed a bug in our code where (in the
happy path) we weren't returning the result from the call to
super()._create(). That was returning None as the object back from a
create call.

Analysis:
Added return to the super call in the overridden version of _create in
Service. Added checks in each of the functional tests to ensure that the
thing that is returned from a create is not None.

Tests:
Ran tests against 12.1.0 for this suite and 11.5.4