Jenkins failure in datacenter.py relating to enabled/disabled
Issues:
Fixes #585

Problem:
This failure occurs due to the way an update is manually handled in the
datacenter. On an update, the enabled/disabled attributes should be
evicted just prior to contacting the device, that way the device reports
back the correct state and updates the object accordingly. If any kwargs
are given to update with enable/disable, then those will be applied
prior to contacting the device as well.

Analysis:
Updated the udpate according to @tar's suggestion, which applies good
logic to evict enabled/disabled from the object's attributes before
updating. Any kwargs with those keys also have the reduce boolean pair
logic applied before contacting the device.

Tests: