During functional testing against 11.5.4, test/functional/tm/ltm/test_virtual.py test fails
Issues:
Fixes #589

Problem:
A profile cannot be created or loaded on 11.5.4 under a virtual without
specifying the partition.

Analysis:
Added a tmos version check in the __init__ of the Profiles resource
under Virtual. If the tmos version is less than 11.6.0, the partition is
required for creation and load.

Tests:
Virtual tests pass against 11.5.4 and 11.6.0