During functional testing against 11.5.4, test/functional/tm/net/test_vlan.py tests fail
Issues:
Fixes #590

Problem:
The tagMode argument or attribute is not allowed on TMOS version 11.5.4
and less. We need to fail in a meaningful way when using that version.

Analysis:
Added a check to create and update in vlan to ensure we raise an
exception if tagMode is present in kwargs or as a self attribute.

Tests:
Tests against 11.6.0 and 11.5.4 pass