Fixes transactions and adds functional tests
Issues:
Fixes #603

Problem:
The transaction classes became non-functional when the create() methods
switched to returning an object instead of modifying the object in place.
Due to a lack of functional testing for transactions, this caused them
to break without any knowledge to us.

Analysis:
This change fixes the transaction classes to handle returned objects instead
of objects that are modified in place. Additionally, it creates a functional
test to prevent this issue from creeping up in the future.

Tests:
tests/functional/tm/transaction/test_transaction.py