Fix broken backwards compat in regards to tmos_version
Issues:
Fixes #638

Problem:
The tmos_version attribute was changed from being a property to being a
key in the _meta_data dictionary on the bigip container. This change
broken backwards compat without use realizing immediately, and we need
to support both methods until we're ready to bump the release to 2.0.0.

Analysis:
Added the @property back to the f5/bigip/__init__.py for tmos_version.
Both ways of accessing tmos_version are now supported

Tests:
Added a test to ensure we can access the tmos_version both ways