Make transaction IDs strings to work with python requests 2.11
Issues:
Fixes #641

Problem:
requests 2.11 changed compatibility to disallow integers inside of headers.
The result was that the transaction functionality in the SDK broke because
transaction IDs are integers and sent in the headers of the request
(X-F5-REST-Coordination-Id).

Analysis:
This patch casts the integers to strings to correct the problem.

Tests:
none