Fixes some 11.6.1 test issues
Issues:
Fixes #658

Problem:
In 11.6.1, some of the resource endpoints gained a partition prefix
and lost a partition attribute. Therefore, those tests would work in
versions prior to 11.6.1, but fail afterwards

Analysis:
The names in version 11.6.1 gained a partition prefix and lost a partition
attribute. The tests need to take this into consideration

Tests:
exiting