Add toggles for modules to symbols file
Issues:
Fixes #660

Problem:
There are cases when you want to run the full test suite, but omit certain
checks based on the lack of modules in your bigip install. There is currently
no way to ignore these tests without getting creative in your test file location
to the py.test command.

Analysis:
This patch adds the ability to specify modules to turn "on" or "off" in a symbols
file when running the funtional tests.

Tests:
none needed