VCMP endpoints need to be added
Issues:
Fixes #666

Problem:
The VCMP endpoints should be added in the tm organizing collection. A
VCMP has a collection of guests. We should be able to retrieve the vlan
and management address for a particular guest as well.

Analysis:
Added the VCMP OrganizingCollection, the Guests Collection, and the
Guest Resource. Added a unit test and some functional tests. The
functional tests can only be run against a vcmp host. So I added a
pytest config option of --vcmp-host, which allows a user to provide an
IP address for a vcmp configured host. Only then will the functional
tests be run.

Tests:
All tests pass against the vcmp host in the Boulder lab