Adds support for validate_only to the transaction
Issues:
Fixes #653

Problem:
The validateOnly functionality of transactions was not exposed to
the end user. Therefore, the user could never know if the transaction
would pass or fail if they ran it without actually running it

Analysis:
This patch adds support to the TransactionContextManager so that
it allows for the validate_only parameter to be provided.

By default, this parameter is set to False to maintain backwards
compatibility.

Tests:
tests/functional/tm/transaction/test_transaction.py::TestTransaction::test_validate_only_true