Adds some more python 2 to 3 fixes
Problem:
This is just one of many python 2 to 3 fixes that need to be made.
This is a progressive enhancement

Analysis:
Various code portions in the resources class and tests were not
python 3 compatible. For instance, the BaseException class deprecated
its "message" attribute in 2.6 and removed it in 3.0. This needed to
be addressed in several areas. Additionally there were some changes
that needed to be made to loops over things that became iterators.

Tests:
f5/bigip/test/test_resource.py