LBaaSv2 adding a tagged vlan not functioning using v12.1 and 11.6.1
Issues:
Fixes #713

Problem:
When making a request to create a vlan, we are adding tagMode as a
required creation parameter if the tagged kwarg was seen in the request.

Analysis:
We removed the update of the required creation parameters. This will now
fail with TagModeDisallowedForTMOSVersion exception if tagMode is in the
request and we are querying against BIG-IP 11.5.4.

Tests:
The tests still pass, indicating we did not break backwards
compatibility.