Addresses v12.1 changes and incompatibilities in GTM sub-module
Fixes Issues: #719,  #679

Updated Files:

    test/functional/gtm/test_pool.py
    test/functional/test_server.py
    f5/bigip/tm/gtm/pool.py
    f5/bigip/tm/gtm/server.py

Updates:

As we started adding Partition to the selflinks in 12.1.0, some tests needed to be amended. There is a serious bug in 12.x branch with members, as described in #719, I have had to create custom create methods for 2 member classes in pools.py.
Corrected subcollection resource names to be singular not plural as per discussion with Tim, amended tests accordingly