ltm policy functional rewrite in tmos version 12.1 needs attention in sdk
Issues:
Fixes #575 #668

Problem:
In 12.1, policies cannot just be created and attached to virtual
servers. There is a workflow to create a draft, then publish, then
attach. For existing policies, you need to create a draft of the policy,
patch the draft, then publish it.

Analysis:
Implemented some defaults to allow the legacy method to execute by
default, that is, a policy is always published with defaults. If the
user is on 12.1.0 and wishes to create a draft then publish. A new
public method exists on the Policy resource.

Tests:
All functional tests pass. These all were used to get coverage in lieu
of unit tests.