Adds ASM Resource class Issues:  #701
Problem:
Second of the ASM PR series, this PR has to be merged after PR #748.

Analysis:
It adds the Asm resource class as well as new meta_data boolean to indicate presence or absence of stats, which is crucial for ASM and future APM module. This PR also has ASM endpoint and 2 classes introduced due to AsmResource testing requirement. Further unit and functional tests for those classes will be introduced in the PR following this one.

Tests:
Unit Tests
Flake8