ltm node sets 'state' as a read-only attribute
Issues:
Fixes #775

Problem:
The 'state' of a node can be modified through REST and is not strictly
read-only. We can remove this and test to ensure it works as expected.
The test will consist of a user creating a node with a specific state,
updating to a specific state, then modifying the state.

Analysis:
Remove 'state' as read-only attribute, then ensured a call to update or
modify does the right thing if state is present and it has a value of
unchecked.

Tests:
Wrote new functional tests to check this behavior