Add functional tests for LTM policy deployments (#789)
* Add functional tests for LTM policy deployments

Issues:
Fixes #788

Problem:
When deploying a policy with the agent, we would like to deploy the
policy along with the rules and conditions and actions in one POST or
PUT. We need a test to assure this operation produces the expected state
on the device.

Analysis:
Added two new test to check this behavior. Found that we can do a PUT on
a policy to change the state of it and the accompanying rules
immediately. We validate that the state is as expected immediately after
updating.

Tests:
Both new tests pass against 12.1 and 11.6

* Updating json load to have proper path to full_policy.json file.

* Added override methods for exists and _load to ensure a rule exists on the device.