Create collection and resource under virtual to allow access to policies (#802)
* Create collection and resource under virtual to allow access to policies

Issues:
Fixes #800

Problem:
The virtual resource currently does not have SDK access to its policies.
Therefore we cannot perform CRUD operations on these policies. We need
to add the collection for Policies_s and the resource Policies.

Analysis:
I had to refactor the policy _create method to make the logic read
better from a developer's viewpoint. We changed some of the underlying
implementation, but we preserved existing functionality. We may have
loosened restrictions for behaviour against older tmos versions. And in
this respect, we have broken backwards compatibility.

Tests:
New tests for virtual, and refactored tests for policy all pass

* Fixing up travis unit test failure in test_resource.py