Fix issue of existence for policies under virtual servers (#804)
* Testing Jenkins run.

* Fix issue of existence for policies under virtual servers`

Issues:
Fixes #805

Problem:
This issue has been encountered before. Here, if we check for existence
or attempt to load a missing policy from a virtual server's namespace,
it will always succeed when using TMOS 11.6.0. Meaning that the GET call
returns a 200 OK, whether the policy exists or not. We need to iterate
of the collection of items from the container and check if the object we
need exists by name.

Analysis:
I commonized the logic to check for existence of a policy under a
virtual to be used by other resources that need it. It is now in the
mixins module.

Tests:
Both policies and rules tests run correctly after being refactor to have
this mixin