Adds remaining parameter parsing for iapps
Issues:
Fixes #807

Problem:
The iapp_parser does a good job of parsing the common formats of
iapps but leaves out some of the more edge cases.

Analysis:
This patch fixes the iapp parser to take into account the remaining
things that can be parsed and reverses some of the expectations
the iapp parser made that were incorrect.

For example, the assumption that requires-modules cannot have a
value of "{ }"

This patch changes the tests that existed because some of those
existing tests were incorrect. Does this mean a backwards incompatible
change?

Tests:
unit tests