Fixes Update method #807
Issues:
Fixes #807

Problem:
The update method strips _meta_data during its update process. This works well when the update suceeds. The problem occurs when we receive exception from bigip. Not only we fail to update (for one reason or the other which is what we intended) but we also leave the calling object with this _meta_data removed.

Analysis:
This patch added the logic to return the object to its original state as well as passing on the raised exception. This means that we do see failure as intended but we no longer have objects with _meta_data stripped.

Tests:
unit tests
flake 8