LTM policy needs to be capable of update and modify with 12.1.0 with legacy
Issues:
Fixes #829

Problem:
There is a legacy key for deploying and modifying policies in 12.1.0.
When set to True, this deploys or modifies a policy as if it were one on
tmos versions prior to 12.1.0. We need to change the modify and update
operations to expect legacy as a keyword argument and allow the user to
update as they see fit.

Analysis:
Made changes to modify and update to allow legacy to be passed in and
used for changes against a 12.1.0 machine. If legacy is passed in and
the tmos version is < 12.1.0, it is simply ignored.

Tests:
All tests against 12.1.0 and 11.6.0 for policies pass