DeviceProvidesIncompatibleKey exception when 'not' is given as object attribute (#837)
* DeviceProvidesIncompatibleKey exception when 'not' is given as object
attribute

Issues:
Fixes #835

Problem:
For LTM Conditions, 'not' is a valid JSON field, and this is a valid
Python keyword. We can get around it (which is what we are doing
currently) by doing a setattr and getattr, but this is not sustainable.
We must brainstorm a few ideas to convert such things that can be
accessible via Python dotted attribute access. (condition._not or
condition.invert).

Analysis:
Implemented a helper function where python keywords are appended with a
trailing underscore. This change takes care of the outgoing request and
the reverse translation on the response.

Tests:
Implemented a few new tests, but unconvered a new bug #836, which
highlights an issue on the device around setting the 'not' attribute.

* Adding logic for creating an ordered dict if the 'check' keyword is seen.