Fixes inability to provide description when creating snat
Issues:
Fixes #863

Problem:
The creation code for snat was too strict when it came to specifying
the allowed arguments.

Analysis:
This patch fixes the creation method to correctly detect that one, and
only one, of the required arguments is provided. This allows one to
provide the remaining optional arguments that a snat can take

Tests:
  * functional