Refactors autogenerated classes
Fixes #503

Problem:
Some of the autogenerated classes had old redundant code and invalid dostrings.

Analysis:
Tidied up the code so it is in line with rest of the SDK. FDB endpoint had Vlan class which was not used, it was removed to avoid confusion for whoever gets to supplement FDB endpoint with missing classes.

Tests:
Functional
Unit
Flake8