Issues: Fixes #933
Problem:
This fixes the fact that the .rpm's and .deb's are not fully tested
prior to being released.

Analysis:
This testing will determine what the dependencies are, install them on a clean
Docker container, and have the packages installed to verify that they are
created properly and will install with the listed dependencies.

Tests:
This is a test in and of itself; however, to properly determine wether
or not this is executing properly would be to run the .travis.yml file
in its typical runtimes.