Issues: Fixes #980
Problem:
There is an issue of creating a means by which to take a source file
that contains all dependencies and incorporate them into the three
building phases performed in any given github repo for F5:

- RPM build (natively)
  - CentOS 7
- DEB build (natively)
  - Ubuntu's "Trusty" 14.04

Analysis:
This introduces a setup_requirements.txt that can be parsed using
pip.req.parse_requirements library method.  This parsing will generate a
generic set of objects that then can be used to derive what modules and
what versions those modules can be at for a proper install.

Tests:
This is tested via the pre-existing .travis.xml package building
architecture.  This scripting is not auto-added to travis because other
commits under this ticket, #980, will handle other aspects of this
automation.