Issues: Fixes #980
Problem:
There is no clean way of getting around the dynanamic passing of
variables throughout the automation of creating the rpm and deb
packages besides hosting a "globals" of sorts.

Analysis:
This is a temporary file generator that stores, in JSON, key locations within
the environment to get around re-discovering these over and over.  Thus,
this allows for different segments of the package-building automation to
be more standalone.

Tests:
This is a part of a test itself; thus, this does not have a test
strictly on its own.  This can be visually tested by running the script
from within the .git/ directory-level of a repo and executing this
script located in the *-dist/scripts/ directory.