Issues: Fixes #980
Problem:
This commit addresses several issues and finishes the list of commits
provided in the fix for issue #980.  The list is as follows:

- The need to have Travis fail for a commit when an .rpm and/or .deb package are
  built that are improper or have improperly named dependencies
- The need to have all package-building materials in a supportable
  format (now in python, not shell script
- The need to have the build scripting materials generic enough to be
  portable to many different github repos and projects (as needed by the
  F5 development team)
- Package and requirements dependency validation/testing

NOTE:
This commit DOES DEPEND on previous commits under this issue, #980.
These commits set up the environment that the built script uses.

Analysis:
This change introduces a build_pkgs.py script and a install_test.py
script.  The Test module houses the logic used to test both the RPM and
DEB packages as needed by first installing the dependencies natively
(using Docker), checking the dependencies for missing items versus the
go-to `setup_requirements.txt` file and then installing the packages.

This introduces changes to the .travis.yml file such that the build
script is called during the installation, which is a critical part of
whether or not Travis passes or fails its test.

Tests:
This script has its own testing structure built in via the
install_test.py.  This script is stand-alone and can be called on its
own from the working directory (just above the .git/ of the given repo).
This script is also, by facto of being a modulino, can also be imported
by a different python module and used to test the packages as a part of
the build script.